Split `Data.Delta` into child modules
